:recycle: [Refactor] eventHandlers로 뺀 함수 main.js 로 롤백 #main
